nightly to stable

+++ menu2/MapActions.gd
@@ -68,29 +68,35 @@ func copy_item_selected(idx:int):
 func item_selected(idx:int):
 	match idx:
 		0: # Delete map
-			Globals.confirm_prompt.open(
-				"Are you sure you want to delete this map? You might not be able to get it back.",
-				"Delete Map",
-				[
-					{ text = "Cancel" },
-					{ text = "OK", wait = 1 }
-				]
-			)
-			Globals.confirm_prompt.s_alert.play()
-			var response:int = yield(Globals.confirm_prompt,"option_selected")
-			Globals.confirm_prompt.close()
-			if response == 1:
-				Globals.confirm_prompt.s_next.play()
-				SSP.selected_song.delete()
-			else:
-				Globals.confirm_prompt.s_back.play()
+			if (
+				(SSP.selected_song.songType == Globals.MAP_SSPM or
+				SSP.selected_song.songType == Globals.MAP_SSPM2) and
+				!SSP.single_map_mode
+			):
+				Globals.confirm_prompt.open(
+					"Are you sure you want to delete this map? You might not be able to get it back.",
+					"Delete Map",
+					[
+						{ text = "Cancel" },
+						{ text = "OK", wait = 1 }
+					]
+				)
+				Globals.confirm_prompt.s_alert.play()
+				var response:int = yield(Globals.confirm_prompt,"option_selected")
+				Globals.confirm_prompt.close()
+				if response == 1:
+					Globals.confirm_prompt.s_next.play()
+					SSP.selected_song.delete()
+				else:
+					Globals.confirm_prompt.s_back.play()
 		1:
 			if !(
 				SSP.selected_song.is_broken or
-				SSP.selected_song.is_builtin or
+				#SSP.selected_song.is_builtin or
 				SSP.selected_song.converted or
 				SSP.selected_song.songType == Globals.MAP_SSPM2 or
-				SSP.selected_song.is_online
+				SSP.selected_song.is_online or
+				SSP.single_map_mode
 			):
 				var res = SSP.selected_song.convert_to_sspm(SSP.selected_song.songType == Globals.MAP_SSPM)
 				if res == "Converted!":
@@ -105,12 +111,13 @@ func item_selected(idx:int):
 				self,
 				"save_song_txt",
 				["*.txt ; Text map data"],
-				"~/Downloads/%s.txt" % [SSP.selected_song.id]
+				#"~/Downloads/%s.txt" % [SSP.selected_song.id]
+				OS.get_system_dir(OS.SYSTEM_DIR_DOWNLOADS) + "/%s.txt" % [SSP.selected_song.id]
 			)
 		5:
 			if !(
 				SSP.selected_song.is_broken or
-				SSP.selected_song.is_builtin or
+				#SSP.selected_song.is_builtin or
 				SSP.selected_song.is_online
 			):
 				audio_data = SSP.selected_song.get_music_buffer()
@@ -120,14 +127,16 @@ func item_selected(idx:int):
 						self,
 						"save_song_audio",
 						["*.mp3 ; mp3 audio file"],
-						"~/Downloads/%s.mp3" % [SSP.selected_song.id]
+						#"~/Downloads/%s.mp3" % [SSP.selected_song.id]
+						OS.get_system_dir(OS.SYSTEM_DIR_DOWNLOADS) + "/%s.mp3" % [SSP.selected_song.id]
 					)
 				elif format == "ogg":
 					Globals.file_sel.save_file(
 						self,
 						"save_song_audio",
 						["*.ogg ; ogg audio file"],
-						"~/Downloads/%s.ogg" % [SSP.selected_song.id]
+						#"~/Downloads/%s.ogg" % [SSP.selected_song.id]
+						OS.get_system_dir(OS.SYSTEM_DIR_DOWNLOADS) + "/%s.ogg" % [SSP.selected_song.id]
 					)
 				else:
 					Globals.notify(Globals.NOTIFY_ERROR,"Unable to determine audio format","Error")
@@ -137,22 +146,25 @@ func item_selected(idx:int):
 func upd(_s=null):
 	visible = true
 	get_popup().set_item_disabled(0,(
-		SSP.selected_song.is_builtin or
+		SSP.single_map_mode or
+		#SSP.selected_song.is_builtin or
 		SSP.selected_song.is_online or !(
 			SSP.selected_song.songType == Globals.MAP_SSPM or
 			SSP.selected_song.songType == Globals.MAP_SSPM2
 		)
 	))
 	get_popup().set_item_disabled(1,(
+		SSP.single_map_mode or
 		SSP.selected_song.is_broken or
-		SSP.selected_song.is_builtin or
+		#SSP.selected_song.is_builtin or
 		SSP.selected_song.converted or
 		SSP.selected_song.songType == Globals.MAP_SSPM2 or
 		SSP.selected_song.is_online
 	))
 	
 	get_popup().set_item_disabled(3,(
-		SSP.selected_song.is_builtin or
+		SSP.single_map_mode or
+		#SP.selected_song.is_builtin or
 		SSP.selected_song.is_online or !(
 			SSP.selected_song.songType == Globals.MAP_SSPM or
 			SSP.selected_song.songType == Globals.MAP_SSPM2
@@ -161,12 +173,12 @@ func upd(_s=null):
 	
 	get_popup().set_item_disabled(5,(
 		SSP.selected_song.is_broken or
-		SSP.selected_song.is_builtin or
+		#SSP.selected_song.is_builtin or
 		SSP.selected_song.is_online
 	))
 	
 	copy_submenu.set_item_disabled(1,(
-		SSP.selected_song.is_builtin or
+		#SSP.selected_song.is_builtin or
 		SSP.selected_song.is_online
 	))
 	
@@ -200,7 +212,7 @@ func _ready():
 	difficulty_submenu.add_radio_check_item("Logic?",4)
 	difficulty_submenu.add_radio_check_item("助 (Tasukete)",5)
 	
-	
+	if SSP.selected_song: upd()
 	SSP.connect("selected_song_changed",self,"upd")
 	yield(get_tree(),"idle_frame")
 	copy_submenu.name = "Copy"
